more int->int16

+++ src/eg3dview.c
@@ -22,7 +22,7 @@ void loadRegion3D() {
 
 // ==== seg000:0x21ca ====
 //
-void render3DView(int camX, int camY, int camZ, long worldX, long worldY, long worldZ, int clipLeft, int clipTop, int clipWidth, int clipHeight) {
+void render3DView(int16 camX, int16 camY, int16 camZ, long worldX, long worldY, long worldZ, int16 clipLeft, int16 clipTop, int16 clipWidth, int16 clipHeight) {
     g_viewParams[7] = clipTop;
     g_viewParams[8] = clipTop + clipHeight - 1;
     g_viewParams[9] = clipLeft;
@@ -32,7 +32,7 @@ void render3DView(int camX, int camY, int camZ, long worldX, long worldY, long w
     *g_viewParams = 0;
     g_viewParams[2] = (unsigned char)((char *)colorLut)[g_skyColorIndex & 0xFF];
     {
-        R3DScene scene = {g_viewParams, camX, camY, camZ, 0, 0, (int)worldZ, 1};
+        R3DScene scene = {g_viewParams, camX, camY, camZ, 0, 0, (int16)worldZ, 1};
         r3d_beginScene(&scene);
     }
     projectObjects(camX, camY, worldX, worldY, worldZ);
@@ -44,7 +44,7 @@ void render3DView(int camX, int camY, int camZ, long worldX, long worldY, long w
 }
 
 // ==== seg000:0x2278 ====
-void waitFrameSync(int frames) {
+void waitFrameSync(int16 frames) {
     uint8 targetTick;
     if (frames > 0) {
         targetTick = (uint8)frames + g_timerTickByte[0];
@@ -53,7 +53,7 @@ void waitFrameSync(int frames) {
 }
 
 // ==== seg000:0x229a loadColorPalette ====
-void loadColorPalette(int idx) {
+void loadColorPalette(int16 idx) {
     memcpy(colorLut, g_colorPalettes + idx * 16, 16);
 }
